feat: LottoStatistics에 수익률을 구하는 기능 구현

+++ src/main/java/lotto/domain/game/LottoRevenueRate.java
@@ -8,7 +8,7 @@ public class LottoRevenueRate {
         this.rate = rate;
     }
 
-    public static LottoRevenueRate of(long revenue, int totalTicketMoney) {
+    public static LottoRevenueRate of(long revenue, long totalTicketMoney) {
         double revenueRate = (double) revenue / totalTicketMoney;
         return new LottoRevenueRate(revenueRate);
     }
